Changes before trying to add notification settings

+++ app/api/rounds/[roundId]/votes/route.ts
@@ -4,6 +4,9 @@ import { getCollection } from "@/lib/mongodb";
 import { Vote } from "@/databaseTypes";
 import { ObjectId } from "mongodb";
 import { triggerRealTimeUpdate } from "@/lib/pusher-server";
+import { getUserLeagues } from "@/lib/data";
+import { getAllRounds } from "@/lib/utils/getAllRounds";
+import { voteNotifications } from "@/lib/notifications";
 
 export async function POST(
   request: NextRequest,
@@ -16,11 +19,37 @@ export async function POST(
     }
 
     const { roundId } = params;
+    if (!roundId) {
+      return NextResponse.json(
+        { error: "Round ID is required" },
+        { status: 400 }
+      );
+    }
+
     const body = await request.json();
+
+    const userLeagues = await getUserLeagues(payload.userId);
+    const { round, league } = (() => {
+      for (const league of userLeagues) {
+        const leagueRounds = getAllRounds(league, {
+          includePending: false,
+          includeFake: false,
+        });
+        for (const round of leagueRounds) {
+          if (round._id === roundId) {
+            return { round, league };
+          }
+        }
+      }
+      return { round: null, league: null };
+    })();
+
+    if (!round) {
+      return NextResponse.json({ error: "Round not found" }, { status: 404 });
+    }
 
     // Get the round and league to validate
     const votesCollection = await getCollection<Vote>("votes");
-
     await votesCollection.deleteMany({ userId: payload.userId, roundId });
 
     const now = Date.now();
@@ -46,6 +75,13 @@ export async function POST(
     await votesCollection.insertMany(data);
 
     triggerRealTimeUpdate();
+    voteNotifications({
+      league,
+      votes: data.map(({ userId }) => ({ userId })),
+      before: {
+        round,
+      },
+    });
 
     return NextResponse.json({ success: true });
   } catch (error) {
